Remove "\n" from end of each line in load_lines
A new line character should not count towards the character limit for
each line. However, `readlines` includes "\n" at the end of each string
as appropriate. Previously, the commit message "Fix Typo." would be
accepted as "Fix Typo.\n" does not end in a ".". Test cases are
modified to reflect the behavior change.